Fix CSS in themes to use points for font sizes, not pixels.
Using physical pixels for font sizes is harmful because it leads
to unreadable interface on high DPI displays.

This change fixes this problem for all 3 themes that use CSS.

Also, maximum size of recent transactions list view in overview tab
has been removed because it was causing this list view to be
so narrow on high DPI displays that transaction amount was clashing
with transaction time if fonts are readable.

Signed-off-by: Oleg Girko <ol@infoserver.lv>