Make CNetAddr and its descendants non-default-constructible.
This is big, ugly and intrusive change. It shows intermediary state of work.

The main focus of this change is abstract away addressing.
CNetAddr and all classes inheriting it
are not default constructible anymore: you must specify network backend to use.

A special DefaultBackend argument is used to specify default TCP backend.
This means that all places that need attention to allow using non-default
network backend, are clearly marked now with DefaultBackend argument.

Also, the fact that CNetAddr and its descendants are not default constructible
breaks serialisation and requires special handling.
This allows to mark all places that need to be changed to allow
serialising addresses with non-default backend.

Also, many places where default constructability of CNetAddr and
its descendants is required for no good reason have been changed.